requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==2.2.3
matplotlib==3.10.1
pillow==11.1.0
streamlit==1.43.2
tqdm
